gun: merge Gun_DrawFlash
