+++ src/tr1/game/gun/gun.c
@@ -5,51 +5,6 @@
 #include <libtrx/config.h>
 #include <libtrx/game/random.h>
 
-void Gun_DrawFlash(LARA_GUN_TYPE weapon_type, CLIP clip)
-{
-    int32_t light;
-    int32_t len;
-    int32_t off;
-
-    switch (weapon_type) {
-    case LGT_MAGNUMS:
-        light = 16 * 256;
-        len = 155;
-        off = 55;
-        break;
-
-    case LGT_UZIS:
-        light = 10 * 256;
-        len = 180;
-        off = 55;
-        break;
-
-    case LGT_SHOTGUN:
-        if (!g_Config.visuals.enable_shotgun_flash) {
-            return;
-        }
-        light = 10 * 256;
-        len = 285;
-        off = 0;
-        break;
-
-    default:
-        light = 20 * 256;
-        len = 155;
-        off = 55;
-        break;
-    }
-
-    Matrix_TranslateRel(0, len, off);
-    Matrix_RotX(-90 * DEG_1);
-    Matrix_RotZ((int16_t)(Random_GetDraw() * 2));
-    Output_CalculateStaticLight(light);
-    const OBJECT *const obj = Object_Get(O_GUN_FLASH);
-    if (obj->loaded) {
-        Object_DrawMesh(obj->mesh_idx, clip, false);
-    }
-}
-
 void Gun_UpdateLaraMeshes(const GAME_OBJECT_ID obj_id)
 {
     const bool lara_has_rifle = Inv_RequestItem(O_SHOTGUN_ITEM);
