+++ src/tr2/game/gun/gun_misc.c
@@ -348,56 +348,3 @@ void Gun_HitTarget(
         }
     }
 }
-
-void Gun_DrawFlash(const LARA_GUN_TYPE weapon_type, const CLIP clip)
-{
-    int16_t shade;
-    int32_t y;
-    int32_t z;
-
-    switch (weapon_type) {
-    case LGT_MAGNUMS:
-        shade = SHADE_NEUTRAL;
-        y = 215;
-        z = 65;
-        break;
-
-    case LGT_UZIS:
-        shade = 2560;
-        y = 200;
-        z = 50;
-        break;
-
-    case LGT_SHOTGUN:
-        return;
-
-    case LGT_M16:
-        Matrix_TranslateRel(0, 400, 99);
-        Matrix_RotX(-85 * DEG_1);
-        Matrix_RotZ(((2 * Random_GetDraw()) & 0x4000) + 0x2000);
-        Output_CalculateStaticLight(2560);
-        Object_DrawMesh(Object_Get(O_M16_FLASH)->mesh_idx, clip, false);
-        return;
-
-    case LGT_FLARE:
-        Matrix_TranslateRel(11, 32, 80);
-        Matrix_RotX(-DEG_90);
-        Matrix_RotY(2 * Random_GetDraw());
-        Output_CalculateStaticLight(2048);
-        Object_DrawMesh(Object_Get(O_FLARE_FIRE)->mesh_idx, clip, false);
-        return;
-
-    default:
-        shade = SHADE_LOW;
-        y = 185;
-        z = 40;
-        break;
-    }
-
-    Matrix_TranslateRel(0, y, z);
-    Matrix_RotX(-DEG_90);
-    Matrix_RotZ(2 * Random_GetDraw());
-    Output_CalculateStaticLight(shade);
-    Object_DrawMesh(Object_Get(O_GUN_FLASH)->mesh_idx, clip, false);
-    return;
-}
